requirements.txt
imgurpython==1.1.6
PyYAML==3.11
requests==2.7.0
twx.botapi==1.0.2
wheel==0.24.0
